scrapers/requirements.txt
requests
beautifulsoup4
python-dotenv
# psycopg2-binary removed (using API now)
